// src/store/common/types.ts
/**
 * This file is part of vue-boilerplate.
 * @link     : https://zhaiyiming.com/
 * @author   : Emil Zhai (root@derzh.com)
 * @modifier : Emil Zhai (root@derzh.com)
 * @copyright: Copyright (c) 2018 TINYMINS.
 */

export enum COMMON {
  SAVE_SCROLL = '@@COMMON/SAVE_SCROLL',
  REDIRECT = '@@COMMON/REDIRECT',
  SHOW_LOADING = '@@COMMON/SHOW_LOADING',
  HIDE_LOADING = '@@COMMON/HIDE_LOADING',
  SHOW_TOAST = '@@COMMON/SHOW_TOAST',
  HIDE_TOAST = '@@COMMON/HIDE_TOAST',
  SHOW_DIALOG = '@@COMMON/SHOW_DIALOG',
  HIDE_DIALOG = '@@COMMON/HIDE_DIALOG',
  SHOW_ACTIONSHEET = '@@COMMON/SHOW_ACTIONSHEET',
  HIDE_ACTIONSHEET = '@@COMMON/HIDE_ACTIONSHEET',
  SHOW_PICKER = '@@COMMON/SHOW_PICKER',
  HIDE_PICKER = '@@COMMON/HIDE_PICKER',
  STORE_INSTANCE = '@@COMMON/STORE_INSTANCE',
  HTTP_INSTANCE = '@@COMMON/HTTP_INSTANCE',
  ROUTER_INSTANCE = '@@COMMON/ROUTER_INSTANCE',
  ENTRY_PARAMS = '@@COMMON/ENTRY_PARAMS',
  ROUTE_BEFORE_CHANGE = '@@COMMON/ROUTE_BEFORE_CHANGE',
  ROUTE_CHANGE = '@@COMMON/ROUTE_CHANGE',
  ROUTE_HISTORY_MODE = '@@COMMON/ROUTE_HISTORY_MODE',
  SET_BODY_SCROLLABLE = '@@COMMON/SET_BODY_SCROLLABLE',
  REMOVE_BODY_SCROLLABLE = '@@COMMON/REMOVE_BODY_SCROLLABLE',
  SET_BODY_AUTO_HEIGHT = '@@COMMON/SET_BODY_AUTO_HEIGHT',
  REMOVE_BODY_AUTO_HEIGHT = '@@COMMON/REMOVE_BODY_AUTO_HEIGHT',
  SET_BODY_BACKGROUND = '@@COMMON/SET_BODY_BACKGROUND',
  REMOVE_BODY_BACKGROUND = '@@COMMON/REMOVE_BODY_BACKGROUND',
  SET_PAGE_TITLE = '@@COMMON/SET_PAGE_TITLE',
  SET_VIEWPORT_SIZE = '@@COMMON/SET_VIEWPORT_SIZE',
  SET_HEADER_HEIGHT = '@@COMMON/SET_HEADER_HEIGHT',
  SET_HEADER_EXTRA_HEIGHT = '@@COMMON/SET_HEADER_EXTRA_HEIGHT',
  REMOVE_HEADER_EXTRA_HEIGHT = '@@COMMON/REMOVE_HEADER_EXTRA_HEIGHT',
  SET_TABBAR_HEIGHT = '@@COMMON/SET_TABBAR_HEIGHT',
  SET_FOOTER_EXTRA_HEIGHT = '@@COMMON/SET_FOOTER_EXTRA_HEIGHT',
  REMOVE_FOOTER_EXTRA_HEIGHT = '@@COMMON/REMOVE_FOOTER_EXTRA_HEIGHT',
  SET_NAVBAR_VISIBLE = '@@COMMON/SET_NAVBAR_VISIBLE',
  REMOVE_NAVBAR_VISIBLE = '@@COMMON/REMOVE_NAVBAR_VISIBLE',
  SET_TABBAR_VISIBLE = '@@COMMON/SET_TABBAR_VISIBLE',
  REMOVE_TABBAR_VISIBLE = '@@COMMON/REMOVE_TABBAR_VISIBLE',
  GET_WECHAT_SDK_INFO = '@@COMMON/GET_WECHAT_SDK_INFO',
  SET_PAGE_SHARE = '@@COMMON/SET_PAGE_SHARE',
}